pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[project]
name = "strands-robots-sim"
dynamic = ["version"]
description = "Robot Control Integration for Strands Agents in Simulated Environments"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Apache-2.0"}
authors = [
    {name = "AWS", email = "opensource@amazon.com"},
]
keywords = ["robots", "agents", "libero", "gym", "strands", "sdk"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "strands-agents",
    "lerobot",
    "lerobot[feetech]",
    "numpy>=1.21.0",
    "torch>=2.0.0",
    "opencv-python-headless>=4.5.0",
    "Pillow>=8.0.0",
    "msgpack>=1.0.0",
    "pyzmq>=27.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-cov",
    "black",
    "isort",
    "flake8",
    "mypy",
]
sim = [
    "libero @ git+https://github.com/Lifelong-Robot-Learning/LIBERO.git",
    "robosuite>=1.4.0",
    "scipy>=1.9.0",
    "mujoco>=2.3.0",
    "gymnasium>=0.26.0",
]

[project.urls]
Homepage = "https://github.com/strands-labs/robots-sim"
Documentation = "https://github.com/strands-labs/robots-sim#readme"
Repository = "https://github.com/strands-labs/robots-sim.git"
Issues = "https://github.com/strands-labs/robots-sim/issues"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.version]
source = "vcs"

[tool.hatch.build.targets.wheel]
packages = ["strands_robots_sim"]

[tool.hatch.envs.default]
dependencies = [
    "pytest>=6.0",
    "pytest-cov",
    "black",
    "isort",
    "flake8",
    "mypy",
]

[tool.hatch.envs.default.scripts]
test = "pytest -m 'not libero' {args:tests}"
lint = ["black --check strands_robots_sim tests", "isort --check-only strands_robots_sim tests", "flake8 strands_robots_sim tests"]
format = ["black strands_robots_sim tests", "isort strands_robots_sim tests"]

[tool.black]
line-length = 120
target-version = ['py310']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 120
multi_line_output = 3

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=strands_robots_sim --cov-report=term-missing"
markers = [
    "mock: tests that use mocking (no external dependencies)",
    "libero: tests that require Libero installation",
]
